fixes, clean up, and new macros for parsing helping.

+++ parsers/indra-astra/src/parser/tokens/expression/literal/primitive/number/mod.rs
@@ -1,5 +1,5 @@
 use crate::parser::tokens::splay_mods;
 
 splay_mods! {
-    number: [integer, decimal]
+    number: [decimal, integer]
 }
